backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "address>=0.1.1",
    "fastapi[standard]>=0.121.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]
